refactor generateDummyData, updateLogs to thunks
