+++ App.js
@@ -10,7 +10,7 @@ import {
   TodayScreen, WeekScreen, ActivityDetailScreen, GoalsScreen, GoalScreen, ActivityFormScreen,
   GoalFormScreen
 } from './src/screens'
-import { store, deleteOneTodaysEntry, upsertTodaysEntry, selectTodayLogByActivityId, selectAllActivities, selectDailyLogById, selectGoalById, createDailyLog, addEntry, createGoal, createActivity, selectTodayLogs } from './src/redux'
+import { store, generateDummyData, updateLogs } from './src/redux'
 
 const Drawer = createDrawerNavigator();
 const Stack = createStackNavigator();
@@ -39,75 +39,10 @@ const GoalsStack = () => (
   </Stack.Navigator>
 )
 
-function generateDummyData(store){
-  store.dispatch(createGoal({name: 'dummy goal'}))
-  store.dispatch(createActivity({name: 'dummy activity', goalId: '0', goal: 'check', repeatMode: 'daily'}))
-  store.dispatch(createActivity({name: 'dummy activity2', goalId: '0', goal: 'check', repeatMode: 'daily'}))
-  const entry = {intervals: [], completed: true, id: 0, archived: true }
-  store.dispatch(createDailyLog({date: DateTime.now()}))
-  store.dispatch(addEntry({date: DateTime.now(), entry}))
-}
-
-function newEntry(activity){
-  return(
-    {
-      intervals: [], 
-      completed: false, 
-      id: activity.id,
-      archived: false
-    }
-  )
-}
-
-function dueToday(activity, activityGoal){
-  const today = DateTime.now()
-  if(!activity.active || !activityGoal.active){
-    return false
-  }
-  if(activity.repeatMode == 'daily'){
-    return true
-  }
-  if(activity.repeatMode == 'select'){
-    if(activity.weekDays[today.weekday]){
-      return true
-    }
-  }
-  return false
-}
-
-function updateLogs(store){
-  const state = store.getState()
-  const today = DateTime.now()
-  
-  if(!selectDailyLogById(state, today)){ 
-    store.dispatch(createDailyLog({date: today}))
-  }
-
-  for(let activity of selectAllActivities(state)){
-    const goal = selectGoalById(state, activity.goalId)
-    const oldLog = selectTodayLogByActivityId(state, activity.id)
-
-    if(dueToday(activity, goal)){
-      if(oldLog){
-        store.dispatch(upsertTodaysEntry({ ...oldLog, archived: false }))
-      }else{
-        const entry = newEntry(activity)
-        store.dispatch(addEntry({date: today, entry}))
-      }
-    }else{
-      if(oldLog?.intervals || oldLog?.completed){
-        store.dispatch(upsertTodaysEntry({ ...oldLog, archived: true }))
-      }else if(oldLog){
-        store.dispatch(deleteOneTodaysEntry(oldLog.id))
-      }
-    }
-  }
-}
-
 export default function App() {
   useEffect(() => {
-    generateDummyData(store)
-    updateLogs(store)
+    store.dispatch(generateDummyData())
+    store.dispatch(updateLogs())
   })
 
   return (
